requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.3